// RockWeb/Obsidian/Elements/HelpBlock.d.ts
import { PropType } from 'vue';
declare const HelpBlock: import("vue").DefineComponent<{
    text: {
        type: PropType<string>;
        required: true;
    };
}, unknown, unknown, {}, {}, import("vue").ComponentOptionsMixin, import("vue").ComponentOptionsMixin, Record<string, any>, string, import("vue").VNodeProps & import("vue").AllowedComponentProps & import("vue").ComponentCustomProps, Readonly<{
    text?: unknown;
} & {
    text: string;
} & {}>, {}>;
export default HelpBlock;